Generalize methods and properties from grid ViewHelpers
